requirements.txt
aiohttp==3.6.2
appnope==0.1.0
arrow==0.10.0
asn1crypto==0.24.0
astroid==2.0.4
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cloudant==2.9.0
cryptography==2.1.4
decorator==4.3.0
dill==0.2.7.1
discord.py==1.3.2
docopt==0.6.2
ecdsa==0.13
enum34==1.1.6
Fabric==1.14.0
funcsigs==1.0.2
gitdb2==2.0.4
GitPython==2.1.9
idna==2.6
ipaddress==1.0.19
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
multidict==4.7.5
multiprocess==0.70.5
nose==1.3.0
numpy==1.17.4
paramiko==2.4.0
parso==0.3.1
pathos==0.2.1
pbr==3.1.1
pep8==1.7.0
pexpect==4.6.0
pickleshare==0.7.4
pox==0.2.3
ppft==1.6.4.7.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.2
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.4.0
pylint==2.1.1
PyNaCl==1.2.1
pyOpenSSL==17.5.0
PySocks==1.6.8
python-dateutil==2.6.1
python-dotenv==0.12.0
pytz==2017.3
PyYAML==5.3.1
releng==1.3.0
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.4
traitlets==4.3.2
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
websockets==8.1
wrapt==1.10.11
yarl==1.4.2
